Arreglar warnings (again) y agregar generador datos ej2

+++ test/SubsidyTest.cpp
@@ -130,7 +130,7 @@ TEST_F(SubsidyTest, deleteOrphanEdge) {
     edges.push_back({1,2,7});
     edges.push_back({2,1,3});
     deleteIsolatedNodes(3, edges);
-    ASSERT_EQ(edges.size(),2);
+    ASSERT_EQ(edges.size(),(size_t)2);
 }
 
 TEST_F(SubsidyTest, isolatedNodes){
